第11次提交：完善了角色管理，现在可以删除角色、修改角色名称、页面上动态添加或移除角色的卡片内容；更换了icon。

+++ src/main/java/com/nxu/service/impl/IdentityServiceImpl.java
@@ -2,9 +2,12 @@ package com.nxu.service.impl;
 
 import com.nxu.entity.Identity;
 import com.nxu.mapper.IdentityMapper;
+import com.nxu.mapper.MenuMapper;
+import com.nxu.mapper.UserMapper;
 import com.nxu.service.IdentityService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -14,23 +17,36 @@ public class IdentityServiceImpl implements IdentityService {
     @Autowired
     private IdentityMapper identityMapper;
 
-    @Override
-    public int insertIdentity(String name) {
-        return identityMapper.insertIdentity(name);
-    }
+    @Autowired
+    private UserMapper userMapper;
+
+    @Autowired
+    private MenuMapper menuMapper;
 
     @Override
-    public int deleteIdentity(int id) {
+    public int addIdentity(Identity identity) {
+        identityMapper.insertIdentity(identity);
+        return identity.getId();
+    }
+
+    @Transactional
+    @Override
+    public int delIdentity(int id) {
+        int i = userMapper.queryUserCount(id);
+        if (i > 0) {
+            return 0;   // 当前角色存在用户，不能删除
+        }
+        menuMapper.deleteRoleMenuByIdentity(id);    // 删除这个角色的所有菜单权限
         return identityMapper.deleteIdentity(id);
     }
 
     @Override
-    public int updateIdentity(String name) {
-        return identityMapper.updateIdentity(name);
+    public int setIdentity(Identity identity) {
+        return identityMapper.updateIdentity(identity);
     }
 
     @Override
-    public List<Identity> selectAllIdentity() {
+    public List<Identity> getAllIdentity() {
         return identityMapper.selectAllIdentity();
     }
 }
